scripts/tecdoc-import/requirements.txt
psycopg2-binary>=2.9.9
mysql-connector-python>=8.2.0
